+++ src/library/DIPS.Mobile.UI/Components/ListItems/ListItem.cs
@@ -1,3 +1,4 @@
+using DIPS.Mobile.UI.Components.ContextMenus;
 using DIPS.Mobile.UI.Components.Dividers;
 using DIPS.Mobile.UI.Components.ListItems.Options;
 using DIPS.Mobile.UI.Effects.Touch;
@@ -256,4 +257,10 @@ public partial class ListItem : ContentView
     private void SetTouchIsEnabled() => Touch.SetIsEnabled(Border, IsEnabled && Command is not null);
 
     private void BindToOptions(ListItemOptions? options) => options?.Bind(this);
+
+    private void AddContextMenu()
+    {
+        ContextMenuEffect.SetMenu(Border, ContextMenu);
+        BindToOptions(ContextMenuOptions);
+    }
 }